Add drawer subfolder
